dom/tests/mochitest/bugs/test_bug545314.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=545314
-->
<head>
  <title>Test for Bug 545314</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=545314">Mozilla Bug 545314</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<iframe id="ifr"></iframe>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 545314 **/
var frame = window.frames[0];
frame.location.assign("http://localhost:8888/");
ok(true, "location.assign didn't throw");

</script>
</pre>
</body>
</html>
